+++ resources/js/common.js
@@ -1,4 +1,5 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
+import 'bootstrap/dist/js/bootstrap.bundle';
 import '@fortawesome/fontawesome-free/js/all.js';
 import '../sass/bootstrap-theming.scss';
 import '../css/common.css';