+++ resources/views/components/menu-button.blade.php
@@ -1,0 +1,7 @@
+<div class="{{$width}} mb-3 mb-md-3">
+    <button class="btn btn-block {{$style}} menu-button" {{$dataToggle ?? ''}}>
+        <span class="menu-text">{{$title}}</span>
+        <br>
+        {{$icon}}
+    </button>
+</div>