+++ resources/views/home.blade.php
@@ -27,9 +27,57 @@
                 </div>
                 <hr class="h2-hr">
                 <div class="row">
-                    @include('menu-buttons.new-borrowing')
-                    @include('menu-buttons.end-borrowing')
-                    @include('menu-buttons.borrowings-history')
+                    @menubutton
+                        @slot('width')
+                            col-md-4
+                        @endslot
+                        @slot('style')
+                            btn-outline-new-borrowing
+                        @endslot
+                        @slot('title')
+                            RÉALISER UN EMPRUNT
+                        @endslot
+                        @slot('dataToggle')
+                            data-toggle="modal" data-target="#newBorrowingModal"
+                        @endslot
+                        @slot('icon')
+                            <span class="fa-layers fa-fw menu-icon">
+                                <i class="fas fa-circle" data-fa-transform="shrink-7 down-5 right-8" data-fa-mask="fas fa-dice"></i>
+                                <i class="fas fa-arrow-up" data-fa-transform="shrink-10 down-5 right-8"></i>
+                            </span>
+                        @endslot
+                    @endmenubutton
+                    @menubutton
+                        @slot('width')
+                            col-md-4
+                        @endslot
+                        @slot('style')
+                            btn-outline-end-borrowing
+                        @endslot
+                        @slot('title')
+                            RÉCUPÉRER UN EMPRUNT
+                        @endslot
+                        @slot('icon')
+                            <span class="fa-layers fa-fw menu-icon">
+                                <i class="fas fa-circle" data-fa-transform="shrink-7 down-5 right-8" data-fa-mask="fas fa-dice"></i>
+                                <i class="fas fa-arrow-down" data-fa-transform="shrink-10 down-5 right-8"></i>
+                            </span>
+                        @endslot
+                    @endmenubutton
+                    @menubutton
+                        @slot('width')
+                            col-md-4
+                        @endslot
+                        @slot('style')
+                            btn-outline-borrowings-history
+                        @endslot
+                        @slot('title')
+                            VOIR L'HISTORIQUE DES EMPRUNTS
+                        @endslot
+                        @slot('icon')
+                            <i class="fas fa-history menu-icon"></i>
+                        @endslot
+                    @endmenubutton
                 </div>
             </div>
         </div>
@@ -40,8 +88,40 @@
                 </div>
                 <hr class="h2-hr">
                 <div class="row">
-                    @include('menu-buttons.view-inventory')
-                    @include('menu-buttons.edit-inventory')
+                    @menubutton
+                        @slot('width')
+                            col-md-6
+                        @endslot
+                        @slot('style')
+                            btn-outline-view-inventory
+                        @endslot
+                        @slot('title')
+                            VOIR L'INVENTAIRE
+                        @endslot
+                        @slot('icon')
+                            <span class="fa-layers fa-fw menu-icon">
+                                <i class="fas fa-circle" data-fa-transform="shrink-7 down-7 right-9" data-fa-mask="fas fa-boxes"></i>
+                                <i class="fas fa-eye" data-fa-transform="shrink-10 down-7 right-9"></i>
+                            </span>
+                        @endslot
+                    @endmenubutton
+                    @menubutton
+                        @slot('width')
+                            col-md-6
+                        @endslot
+                        @slot('style')
+                            btn-outline-edit-inventory
+                        @endslot
+                        @slot('title')
+                            GÉRER L'INVENTAIRE
+                        @endslot
+                        @slot('icon')
+                             <span class="fa-layers fa-fw menu-icon">
+                                <i class="fas fa-circle" data-fa-transform="shrink-7 down-7 right-9" data-fa-mask="fas fa-boxes"></i>
+                                <i class="fas fa-wrench" data-fa-transform="shrink-10 down-7 right-9"></i>
+                            </span>
+                        @endslot
+                    @endmenubutton
                 </div>
             </div>
         </div>
@@ -52,8 +132,40 @@
                 </div>
                 <hr class="h2-hr">
                 <div class="row">
-                    @include('menu-buttons.admin')
-                    @include('menu-buttons.lender')
+                    @menubutton
+                        @slot('width')
+                            col-md-6
+                        @endslot
+                        @slot('style')
+                            btn-outline-admin
+                        @endslot
+                        @slot('title')
+                            GÉRER LES ADMINISTRATEURS
+                        @endslot
+                        @slot('icon')
+                            <span class="fa-layers fa-fw menu-icon">
+                                <i class="fas fa-circle" data-fa-transform="shrink-7 down-6 right-6" data-fa-mask="fas fa-users"></i>
+                                <i class="fas fa-crown" data-fa-transform="shrink-10 down-6 right-6"></i>
+                            </span>
+                        @endslot
+                    @endmenubutton
+                    @menubutton
+                        @slot('width')
+                            col-md-6
+                        @endslot
+                        @slot('style')
+                            btn-outline-lender
+                        @endslot
+                        @slot('title')
+                            GÉRER LES PRÊTEURS
+                        @endslot
+                        @slot('icon')
+                             <span class="fa-layers fa-fw menu-icon">
+                                <i class="fas fa-circle" data-fa-transform="shrink-7 down-6 right-6" data-fa-mask="fas fa-users"></i>
+                                <i class="fas fa-dice" data-fa-transform="shrink-10 down-6 right-6"></i>
+                             </span>
+                        @endslot
+                    @endmenubutton
                 </div>
             </div>
         </div>