+++ changelog.md
@@ -5,12 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.5.0] - 2018-09-05
+## [0.5.0] - Unreleased
 
 ### Added
 
-* Main menu
-    * Modularity for buttons creation.
+* Views
+    * Blade components for menu buttons and modals.
+* Public
+    * CSS for input fields in modals.
 
 ### Changed